8-application-demos/6-kalshi-bet-predictor/requirements.txt
accelerate==1.10.1
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
distro==1.9.0
exa-py==1.16.1
filelock==3.20.0
fsspec==2025.9.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.11
Jinja2==3.1.6
jiter==0.11.1
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
openai==2.5.0
packaging==25.0
playwright==1.55.0
psutil==7.1.0
pydantic==2.12.3
pydantic_core==2.41.4
pyee==13.0.0
PyPDF2==3.0.1
python-dotenv==1.1.1
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
sniffio==1.3.1
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
